[REFACTOR] refactor Video Conference main application
this refactoring was done to simplify the logic in joining the conference and previewing the user's input/output devices. Instead of using a seperate condition for when the user is already out of the landing page or not, we now just check whether the user isInCall or isInPreview. We've also removed duplicate logic in the joining procedure to ensure consistency and maintainability