[RED] create test for date formatter util
This utility function will be used to format the dates mainly in the response test application. In order to pass these tests, the dates need to become properly formatted according to these criterias: 1. for dates that has a difference of less than a day than the current relative date will be displayed as <HH:MM> (AM | PM), 2. dates that has a difference of more than a day(date difference not 24 hour difference) or less than a year will be displayed as <MM DD>, 3. finally for dates seperated by more than a year it will be displayed as <DD/MM/YY>