[REFACTOR] refactor Window component
Complete overhaul of the window component. We have architected a better model of updating the state tied to each window. The window will now only render once when being dragged and resized. We now rely more on CSS rather than state since it produces better predictability. During resizing we have chosen to opt-out of react's render cycle and directly manipulate the DOM rather than the state. This allows us to only re-render once to update the state of the window on resize end. The new window component is now ~5-10x faster based on the profiling done with React devtools.