[RED] create stubs and interfaces for useUpload hook
this hook will aid in uploading files to the server. Why not just use the usePostRequest hook? Well, this hook first of all sends it's data using the multipart/form-data format which is more efficient at sending files over the wire since it takes 33% less memory compared to JSON with base64 strings. Secondly, this hook uses the old XMLHttpRequest API that allow us to monitor the upload progress of our request, this is useful information to the user since it enables them to gauge how long until their file is uploaded.