[GREEN] implement useUpload hook
the use upload hook was tested in 3 main configurations. This implementation must ensure that it will upload correctly first. Then we made sure that even if there was an error on the server, the hook will graciously fallback to it's error state which is why we provided callbacks such as onUploadError. Finally we had to make sure that it can handle auth errors and refresh tokens automatically when necessary