[REFACTOR] fix infinite loading bug during authentication on first load
There was a bug where if the user left the website authenticated and comes back to a migrated/fresh database, the authentication provider would infinitely try to query for the user's info eventhough they don't exist anymore in the database. This was caused by the fact that the locally saved tokens on the user's device were still valid, although they did not map to any user when queried to the backend. Since the tokens were still valid we had the provider setup in a way where it was optimistic that it WILL get user info every time it requested it with valid tokens. Thus we had to make it where, even though we have valid tokens, if the user info request fails, we log the user out.